moved CorgiRate from CorgiPhotos folder to corgi
