+++ src/modules/common/corgi/CorgiPhotos/index.js
@@ -2,7 +2,6 @@ import CorgiCommon from './CorgiCommon';
 import CorgiUncommon from './CorgiUncommon';
 import CorgiRare from './CorgiRare';
 import CorgiVeryRare from './CorgiVeryRare';
-import CorgiRate from './CorgiRate';
 import SwitchCorgiPhoto from './SwitchCorgiPhoto';
 
-export { CorgiCommon, CorgiUncommon, CorgiRare, CorgiVeryRare, CorgiRate, SwitchCorgiPhoto };
+export { CorgiCommon, CorgiUncommon, CorgiRare, CorgiVeryRare, SwitchCorgiPhoto };
