+++ src/modules/common/corgi/CorgiRate/CorgiRate.scss
@@ -1,0 +1,20 @@
+@import 'mixins';
+@import 'classes';
+
+.corgi-rate {
+  max-width: 15rem;
+
+  &__title {
+    margin-bottom: 0.5rem;
+
+    font-weight: 500;
+  }
+
+  &__list {
+    @include dis-flex-column(flex-start, stretch);
+  }
+
+  &__item {
+    @include margin-between(0.5rem, column);
+  }
+}
