+++ src/modules/common/corgi/CorgiRate/index.js
@@ -1,0 +1,3 @@
+import CorgiRate from './CorgiRate';
+
+export default CorgiRate;
